chapter-2/requirements.txt
# PPT 생성 엔진 의존성
python-pptx>=0.6.21
Pillow>=10.0.0
